feat: Add TelegramBotSend task
I removed dependency on external lib to send message to the Telegram bot
API because I only really needed to be able to send message.

I made a really naive sendMessage that use URL.openConnection. I thought
it could be nice to inject a micronaut httpclient but could not make
that work in an acceptable time. Also I did not find a solution to
prevent the Kestra documentation showing the internal client when I did
that.